+++ code/src/cljs/sixsq/nuvla/ui/deployment_dialog/views.cljs
@@ -39,7 +39,7 @@
   [{:keys [step-id] :as step-state}]
   (let [cred-id                (subscribe [::subs/selected-credential-id])
         credential-loading?    (subscribe [::creds-subs/credential-check-loading? @cred-id])
-        credential-valid?    (subscribe [::creds-subs/credential-check-status-valid? @cred-id])
+        credential-valid?      (subscribe [::creds-subs/credential-check-status-valid? @cred-id])
         check-status           (creds-utils/credential-check-status
                                  @credential-loading? (not @credential-valid?))
         credentials-completed? (subscribe [::subs/credentials-completed?])]
@@ -137,7 +137,8 @@
         button-color     (subscribe [::subs/modal-action-button-color])
         button-disabled? (subscribe [::subs/modal-action-button-disabled?])
         operation        (subscribe [::subs/modal-operation])
-        exec-mode        (subscribe [::subs/execution-mode])]
+        exec-mode        (subscribe [::subs/execution-mode])
+        submit-loading?  (subscribe [::subs/submit-loading?])]
     (fn [show-data?]
       (let [hide-fn   #(do
                          (when (= (:state @deployment) "CREATED")
@@ -174,6 +175,7 @@
          [ui/ModalActions
           [ui/Button {:color    @button-color
                       :disabled @button-disabled?
+                      :loading  @submit-loading?
                       :on-click submit-fn}
            [ui/Icon {:name @button-icon}]
            @button-text]]]))))